[issue/12-add_news] move api method to postapicontroller
